+++ MapboxSearch/autofill/src/main/java/com/mapbox/search/autofill/AddressAutofillImpl.kt
@@ -1,6 +1,10 @@
 package com.mapbox.search.autofill
 
+import com.mapbox.bindgen.Expected
+import com.mapbox.bindgen.ExpectedFactory.createError
+import com.mapbox.bindgen.ExpectedFactory.createValue
 import com.mapbox.geojson.Point
+import com.mapbox.search.base.BaseResponseInfo
 import com.mapbox.search.base.core.createCoreReverseGeoOptions
 import com.mapbox.search.base.core.createCoreSearchOptions
 import com.mapbox.search.base.result.BaseSearchResult
@@ -10,27 +14,44 @@ import kotlinx.coroutines.Deferred
 import kotlinx.coroutines.async
 import kotlinx.coroutines.coroutineScope
 
+internal sealed class SearchSelectionResponse {
+
+    data class Suggestions(
+        val suggestions: List<BaseSearchSuggestion>,
+        val responseInfo: BaseResponseInfo,
+    ) : SearchSelectionResponse()
+
+    data class Result(
+        val suggestion: BaseSearchSuggestion,
+        val result: BaseSearchResult,
+        val responseInfo: BaseResponseInfo,
+    ) : SearchSelectionResponse()
+
+    data class CategoryResult(
+        val suggestion: BaseSearchSuggestion,
+        val results: List<BaseSearchResult>,
+        val responseInfo: BaseResponseInfo,
+    ) : SearchSelectionResponse()
+}
+
 /**
  * Temporary implementation of the [AddressAutofill] based on the two-step search.
  */
 internal class AddressAutofillImpl(private val searchEngine: AutofillSearchEngine) : AddressAutofill {
 
-    override suspend fun suggestions(point: Point, options: AddressAutofillOptions): AddressAutofillResponse {
+    override suspend fun suggestions(point: Point, options: AddressAutofillOptions): Expected<Exception, List<AddressAutofillSuggestion>> {
         val coreOptions = createCoreReverseGeoOptions(
             point = point,
             countries = options.countries?.map { it.code },
             language = options.language?.let { listOf(it.code) },
         )
 
-        return when (val response = searchEngine.search(coreOptions)) {
-            is SearchResultsResponse.Results -> AddressAutofillResponse.Suggestions(
-                response.results.toAddressAutofillSuggestions()
-            )
-            is SearchResultsResponse.Error -> AddressAutofillResponse.Error(response.e)
+        return searchEngine.search(coreOptions).mapValue { (results, _) ->
+            results.toAddressAutofillSuggestions()
         }
     }
 
-    override suspend fun suggestions(query: Query, options: AddressAutofillOptions): AddressAutofillResponse {
+    override suspend fun suggestions(query: Query, options: AddressAutofillOptions): Expected<Exception, List<AddressAutofillSuggestion>> {
         val response = searchEngine.search(
             query = query.query,
             options = createCoreSearchOptions(
@@ -41,30 +62,26 @@ internal class AddressAutofillImpl(private val searchEngine: AutofillSearchEngin
             )
         )
 
-        return when (response) {
-            is SearchSuggestionsResponse.Suggestions -> forwardGeocoding(response.suggestions)
-            is SearchSuggestionsResponse.Error -> AddressAutofillResponse.Error(response.e)
+        return if (response.isValue) {
+            forwardGeocoding(requireNotNull(response.value).first)
+        } else {
+            createError(requireNotNull(response.error))
         }
     }
 
-    private suspend fun forwardGeocoding(suggestions: List<BaseSearchSuggestion>): AddressAutofillResponse {
+    private suspend fun forwardGeocoding(suggestions: List<BaseSearchSuggestion>): Expected<Exception, List<AddressAutofillSuggestion>> {
         return when {
             suggestions.isEmpty() -> {
-                AddressAutofillResponse.Suggestions(emptyList())
+                createValue(emptyList())
             }
             suggestions.all { it.isBatchResolveSupported } -> {
-                when (val selectionResponse = searchEngine.select(suggestions)) {
-                    is SearchMultipleSelectionResponse.Results -> AddressAutofillResponse.Suggestions(
-                        selectionResponse.results.toAddressAutofillSuggestions()
-                    )
-                    is SearchMultipleSelectionResponse.Error -> AddressAutofillResponse.Error(
-                        selectionResponse.e
-                    )
+                searchEngine.select(suggestions).mapValue { (_, results, _) ->
+                    results.toAddressAutofillSuggestions()
                 }
             }
             else -> {
                 coroutineScope {
-                    val deferred: List<Deferred<SearchSelectionResponse>> = suggestions
+                    val deferred: List<Deferred<Expected<Exception, SearchSelectionResponse>>> = suggestions
                         // Filtering in order to avoid infinite recursion
                         // because of some specific suggestions like "Did you mean recursion?"
                         .filter { it.type !is BaseSearchSuggestionType.Query }
@@ -76,40 +93,40 @@ internal class AddressAutofillImpl(private val searchEngine: AutofillSearchEngin
 
                     val selectionResponses = deferred.map { it.await() }
 
-                    val responses: List<AddressAutofillResponse> = selectionResponses
-                        .map { selectionResponse ->
-                            when (selectionResponse) {
-                                is SearchSelectionResponse.Suggestions -> {
-                                    forwardGeocoding(selectionResponse.suggestions)
-                                }
-                                is SearchSelectionResponse.Result -> {
-                                    val autofillSuggestion = selectionResponse.result.toAddressAutofillSuggestion()
-                                    if (autofillSuggestion != null) {
-                                        AddressAutofillResponse.Suggestions(listOf(autofillSuggestion))
-                                    } else {
-                                        AddressAutofillResponse.Suggestions(emptyList())
+                    val responses: List<Expected<Exception, List<AddressAutofillSuggestion>>> = selectionResponses
+                        .map { result ->
+                            if (result.isValue) {
+                                when (val response = requireNotNull(result.value)) {
+                                    is SearchSelectionResponse.Suggestions -> {
+                                        forwardGeocoding(response.suggestions)
+                                    }
+                                    is SearchSelectionResponse.Result -> {
+                                        val autofillSuggestion = response.result.toAddressAutofillSuggestion()
+                                        if (autofillSuggestion != null) {
+                                            createValue(listOf(autofillSuggestion))
+                                        } else {
+                                            createValue(emptyList())
+                                        }
+                                    }
+                                    is SearchSelectionResponse.CategoryResult -> {
+                                        createValue(response.results.toAddressAutofillSuggestions())
                                     }
                                 }
-                                is SearchSelectionResponse.CategoryResult -> {
-                                    AddressAutofillResponse.Suggestions(selectionResponse.results.toAddressAutofillSuggestions())
-                                }
-                                is SearchSelectionResponse.Error -> {
-                                    AddressAutofillResponse.Error(selectionResponse.e)
-                                }
+                            } else {
+                                createError(requireNotNull(result.error))
                             }
                         }
 
                     // If at least one response completed successfully, return it.
-                    if (responses.isNotEmpty() && responses.all { it is AddressAutofillResponse.Error }) {
+                    if (responses.isNotEmpty() && responses.all { it.isError }) {
                         responses.first()
                     } else {
                         responses.asSequence()
-                            .mapNotNull { it as? AddressAutofillResponse.Suggestions }
-                            .map { it.suggestions }
+                            .mapNotNull { it.value }
                             .flatten()
                             .toList()
                             .let {
-                                AddressAutofillResponse.Suggestions(it)
+                                createValue(it)
                             }
                     }
                 }
@@ -123,7 +140,7 @@ internal class AddressAutofillImpl(private val searchEngine: AutofillSearchEngin
 
         fun BaseSearchResult.toAddressAutofillSuggestion(): AddressAutofillSuggestion? {
             // Filtering incomplete results
-            val autofillAddress = AddressComponents.fromCoreSdkAddress(address) ?: return null
+            val autofillAddress = AddressComponents.fromCoreSdkAddress(address, metadata) ?: return null
             val validCoordinate = coordinate
 
             return AddressAutofillSuggestion(