+++ MapboxSearch/autofill/src/test/java/com/mapbox/search/autofill/AddressAutofillSuggestionTest.kt
@@ -7,61 +7,51 @@ import com.mapbox.search.common.tests.ReflectionObjectsFactory
 import com.mapbox.search.common.tests.ToStringVerifier
 import com.mapbox.search.common.withPrefabTestBoundingBox
 import com.mapbox.search.common.withPrefabTestPoint
-import com.mapbox.test.dsl.TestCase
+import io.mockk.mockk
 import nl.jqno.equalsverifier.EqualsVerifier
-import org.junit.jupiter.api.TestFactory
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Test
 
 internal class AddressAutofillSuggestionTest {
 
-    @TestFactory
-    fun `Check AddressAutofillSuggestion result() function`() = TestCase {
-        Given("${AddressAutofillSuggestion::class.java.simpleName} instance") {
-            val name = "Test name"
-            val formattedAddress = "Test formatted address"
-            val coordinate = Point.fromLngLat(10.0, 11.0)
-            val address = AddressComponents.fromCoreSdkAddress(BaseSearchAddress(country = "test"))!!
+    @Test
+    fun `Check AddressAutofillSuggestion result() function`() {
+        val address = requireNotNull(
+            AddressComponents.fromCoreSdkAddress(
+                BaseSearchAddress(country = "test"),
+                mockk()
+            )
+        )
 
-            val suggestion = AddressAutofillSuggestion(name, formattedAddress, coordinate, address)
+        val suggestion = AddressAutofillSuggestion(
+            name = "Test name",
+            formattedAddress = "Test formatted address",
+            coordinate = Point.fromLngLat(10.0, 11.0),
+            address = address
+        )
 
-            When("result() function called") {
-                val result = AddressAutofillResult(suggestion, address)
-
-                Then("Returned value should be $result",
-                    suggestion.result(),
-                    result
-                )
-            }
-        }
+        assertEquals(
+            AddressAutofillResult(suggestion, address),
+            suggestion.result()
+        )
     }
 
-    @TestFactory
-    fun `Check AddressAutofillSuggestion's equals() and hashCode()`() = TestCase {
-        Given("${AddressAutofillSuggestion::class.java.simpleName} class") {
-            When("Call equals() and hashCode()") {
-                Then("equals() and hashCode() should be implemented correctly") {
-                    EqualsVerifier.forClass(AddressAutofillSuggestion::class.java)
-                        .withPrefabTestPoint()
-                        .withPrefabTestBoundingBox()
-                        .verify()
-                }
-            }
-        }
+    @Test
+    fun `equals() and hashCode() functions are correct`() {
+        EqualsVerifier.forClass(AddressAutofillSuggestion::class.java)
+            .withPrefabTestPoint()
+            .withPrefabTestBoundingBox()
+            .verify()
     }
 
-    @TestFactory
-    fun `Check AddressAutofillSuggestion's toString()`() = TestCase {
-        Given("${AddressAutofillSuggestion::class.java.simpleName} class") {
-            When("Call toString()") {
-                Then("toString() function should be implemented correctly") {
-                    ToStringVerifier(
-                        clazz = AddressAutofillSuggestion::class,
-                        objectsFactory = ReflectionObjectsFactory(
-                            extraCreators = CommonSdkTypeObjectCreators.ALL_CREATORS
-                        ),
-                        ignoredProperties = listOf("address"),
-                    ).verify()
-                }
-            }
-        }
+    @Test
+    fun `toString() function is correct`() {
+        ToStringVerifier(
+            clazz = AddressAutofillSuggestion::class,
+            objectsFactory = ReflectionObjectsFactory(
+                extraCreators = CommonSdkTypeObjectCreators.ALL_CREATORS
+            ),
+            ignoredProperties = listOf("address"),
+        ).verify()
     }
 }