+++ MapboxSearch/autofill/src/test/java/com/mapbox/search/autofill/AddressComponentsTest.kt
@@ -1,169 +1,98 @@
 package com.mapbox.search.autofill
 
+import com.mapbox.search.base.core.countryIso1
+import com.mapbox.search.base.core.countryIso2
 import com.mapbox.search.base.result.BaseSearchAddress
 import com.mapbox.search.common.CommonSdkTypeObjectCreators
+import com.mapbox.search.common.createTestResultMetadata
 import com.mapbox.search.common.tests.ReflectionObjectsFactory
 import com.mapbox.search.common.tests.ToStringVerifier
 import com.mapbox.search.common.withPrefabTestBoundingBox
 import com.mapbox.search.common.withPrefabTestPoint
-import com.mapbox.test.dsl.TestCase
 import io.mockk.every
 import io.mockk.mockk
 import nl.jqno.equalsverifier.EqualsVerifier
-import org.junit.jupiter.api.TestFactory
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Assertions.assertNotNull
+import org.junit.jupiter.api.Assertions.assertNull
+import org.junit.jupiter.api.Test
 
 internal class AddressComponentsTest {
 
-    @TestFactory
-    fun `Check AddressComponents properties`() = TestCase {
-        Given("${AddressComponents::class.java.simpleName} class") {
-            When("All AddressComponents properties accessed") {
-                val searchAddress = BaseSearchAddress(
-                    houseNumber = "5",
-                    street = "Rue De Marseille",
-                    neighborhood = "Porte-Saint-Martin",
-                    locality = "10th arrondissement of Paris",
-                    postcode = "75010",
-                    place = "Paris",
-                    district = "Paris district",
-                    region = "Paris region",
-                    country = "France"
-                )
+    @Test
+    fun `Check AddressComponents properties`() {
+        val searchAddress = BaseSearchAddress(
+            houseNumber = "5",
+            street = "Rue De Marseille",
+            neighborhood = "Porte-Saint-Martin",
+            locality = "10th arrondissement of Paris",
+            postcode = "75010",
+            place = "Paris",
+            district = "Paris district",
+            region = "Paris region",
+            country = "France"
+        )
 
-                val addressComponents = requireNotNull(AddressComponents.fromCoreSdkAddress(searchAddress))
+        val coreMetadata = createTestResultMetadata(
+            data = hashMapOf("iso_3166_1" to "fra", "iso_3166_2" to "fr")
+        )
 
-                Then(
-                    "houseNumber should be retrieved from SearchAddress",
-                    searchAddress.houseNumber,
-                    addressComponents.houseNumber
-                )
+        val addressComponents = requireNotNull(
+            AddressComponents.fromCoreSdkAddress(searchAddress, coreMetadata)
+        )
 
-                Then(
-                    "street should be retrieved from SearchAddress",
-                    searchAddress.street,
-                    addressComponents.street
-                )
-
-                Then(
-                    "neighborhood should be retrieved from SearchAddress",
-                    searchAddress.neighborhood,
-                    addressComponents.neighborhood
-                )
-
-                Then(
-                    "locality should be retrieved from SearchAddress",
-                    searchAddress.locality,
-                    addressComponents.locality
-                )
-
-                Then(
-                    "postcode should be retrieved from SearchAddress",
-                    searchAddress.postcode,
-                    addressComponents.postcode
-                )
-
-                Then(
-                    "place should be retrieved from SearchAddress",
-                    searchAddress.place,
-                    addressComponents.place
-                )
-
-                Then(
-                    "district should be retrieved from SearchAddress",
-                    searchAddress.district,
-                    addressComponents.district
-                )
-
-                Then(
-                    "region should be retrieved from SearchAddress",
-                    searchAddress.region,
-                    addressComponents.region
-                )
-
-                Then(
-                    "country should be retrieved from SearchAddress",
-                    searchAddress.country,
-                    addressComponents.country
-                )
-            }
-        }
+        assertEquals(searchAddress.houseNumber, addressComponents.houseNumber)
+        assertEquals(searchAddress.street, addressComponents.street)
+        assertEquals(searchAddress.neighborhood, addressComponents.neighborhood)
+        assertEquals(searchAddress.locality, addressComponents.locality)
+        assertEquals(searchAddress.postcode, addressComponents.postcode)
+        assertEquals(searchAddress.place, addressComponents.place)
+        assertEquals(searchAddress.district, addressComponents.district)
+        assertEquals(searchAddress.region, addressComponents.region)
+        assertEquals(searchAddress.country, addressComponents.country)
+        assertEquals(coreMetadata.countryIso1, addressComponents.countryIso1)
+        assertEquals(coreMetadata.countryIso2, addressComponents.countryIso2)
     }
 
-    @TestFactory
-    fun `Check AddressComponents fromCoreSdkAddress() function`() = TestCase {
-        Given("${AddressComponents::class.java.simpleName} class") {
-            When("fromCoreSdkAddress() called with non empty SearchAddress") {
-                val searchAddress = BaseSearchAddress(country = "France")
-
-                Then(
-                    "fromCoreSdkAddress() should return non null",
-                    false,
-                    AddressComponents.fromCoreSdkAddress(searchAddress) == null
-                )
-            }
-
-            When("fromCoreSdkAddress() called with empty SearchAddress") {
-                val emptySearchAddress = BaseSearchAddress()
-
-                Then(
-                    "fromCoreSdkAddress() should return null",
-                    null,
-                    AddressComponents.fromCoreSdkAddress(emptySearchAddress)
-                )
-            }
-        }
+    @Test
+    fun `Check AddressComponents fromCoreSdkAddress() function`() {
+        assertNotNull(AddressComponents.fromCoreSdkAddress(BaseSearchAddress(country = "France"), mockk()))
+        assertNull(AddressComponents.fromCoreSdkAddress(BaseSearchAddress(), mockk()))
     }
 
-    @TestFactory
-    fun `Check AddressComponents formattedAddress() function`() = TestCase {
-        Given("${AddressComponents::class.java.simpleName} class") {
-            When("formattedAddress() called") {
-                val formattedAddress = "Rue de Marseille, Paris, France"
+    @Test
+    fun `Check AddressComponents formattedAddress() function`() {
+        val formattedAddress = "Rue de Marseille, Paris, France"
 
-                val searchAddress = mockk<BaseSearchAddress>(relaxed = true)
-                every { searchAddress.street } returns "Rue de Marseille"
-                every { searchAddress.place } returns "Paris"
-                every { searchAddress.country } returns "France"
+        val searchAddress = mockk<BaseSearchAddress>(relaxed = true)
+        every { searchAddress.street } returns "Rue de Marseille"
+        every { searchAddress.place } returns "Paris"
+        every { searchAddress.country } returns "France"
 
-                val addressComponents = requireNotNull(AddressComponents.fromCoreSdkAddress(searchAddress))
+        val addressComponents = requireNotNull(AddressComponents.fromCoreSdkAddress(searchAddress, mockk()))
 
-                Then(
-                    "formattedAddress() should return $formattedAddress",
-                    formattedAddress,
-                    addressComponents.formattedAddress()
-                )
-            }
-        }
+        assertEquals(
+            formattedAddress,
+            addressComponents.formattedAddress()
+        )
     }
 
-    @TestFactory
-    fun `Check AddressComponents equals() and hashCode()`() = TestCase {
-        Given("${AddressComponents::class.java.simpleName} class") {
-            When("Call equals() and hashCode()") {
-                Then("equals() and hashCode() should be implemented correctly") {
-                    EqualsVerifier.forClass(AddressComponents::class.java)
-                        .withPrefabTestPoint()
-                        .withPrefabTestBoundingBox()
-                        .verify()
-                }
-            }
-        }
+    @Test
+    fun `equals() and hashCode() functions are correct`() {
+        EqualsVerifier.forClass(AddressComponents::class.java)
+            .withPrefabTestPoint()
+            .withPrefabTestBoundingBox()
+            .verify()
     }
 
-    @TestFactory
-    fun `Check AddressComponents toString()`() = TestCase {
-        Given("${AddressComponents::class.java.simpleName} class") {
-            When("Call toString()") {
-                Then("toString() function should be implemented correctly") {
-                    ToStringVerifier(
-                        clazz = AddressComponents::class,
-                        objectsFactory = ReflectionObjectsFactory(
-                            extraCreators = CommonSdkTypeObjectCreators.ALL_CREATORS
-                        ),
-                        ignoredProperties = listOf("coreSdkAddress"),
-                    ).verify()
-                }
-            }
-        }
+    @Test
+    fun `toString() function is correct`() {
+        ToStringVerifier(
+            clazz = AddressComponents::class,
+            objectsFactory = ReflectionObjectsFactory(
+                extraCreators = CommonSdkTypeObjectCreators.ALL_CREATORS
+            ),
+            ignoredProperties = listOf("coreSdkAddress", "coreMetadata"),
+        ).verify()
     }
 }