+++ MapboxSearch/sample/src/main/java/com/mapbox/search/sample/api/AddressAutofillKotlinExampleActivity.kt
@@ -6,7 +6,6 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.lifecycleScope
 import com.mapbox.search.autofill.AddressAutofill
 import com.mapbox.search.autofill.AddressAutofillOptions
-import com.mapbox.search.autofill.AddressAutofillResponse
 import com.mapbox.search.autofill.Query
 import com.mapbox.search.sample.R
 
@@ -24,18 +23,13 @@ class AddressAutofillKotlinExampleActivity : AppCompatActivity() {
         lifecycleScope.launchWhenCreated {
             val query = Query.create("740 15th St NW, Washington") ?: return@launchWhenCreated
 
-            val response = addressAutofill.suggestions(
+            addressAutofill.suggestions(
                 query = query,
                 options = AddressAutofillOptions()
-            )
-
-            when (response) {
-                is AddressAutofillResponse.Suggestions -> {
-                    Log.i("SearchApiExample", "Autofill suggestions: ${response.suggestions}")
-                }
-                is AddressAutofillResponse.Error -> {
-                    Log.i("SearchApiExample", "Autofill suggestions error", response.error)
-                }
+            ).onValue { suggestions ->
+                Log.i("SearchApiExample", "Autofill suggestions: $suggestions")
+            }.onError { error ->
+                Log.i("SearchApiExample", "Autofill suggestions error", error)
             }
         }
     }