+++ CHANGELOG.md
@@ -4,12 +4,14 @@
 
 ### New features
 - [Discover API] Discover API preview is available.
+- [Autofill] Now `AddressComponents` provides new properties `AddressComponents.countryIso1` and `AddressComponents.countryIso2`.
 
 ### Breaking changes
 - [CORE] `OpenHours`, `OpenPeriod`, `WeekDay`, `WeekTimestamp`, `ParkingData` classes have been moved to `com.mapbox.search.common.metadata` package.
 - [CORE] `Language` has been renamed to `IsoLanguageCode` and moved to `com.mapbox.search.common` package.
 - [CORE] `Country` has been renamed to `IsoCountryCode` and moved to `com.mapbox.search.common` package.
 - [Autofill] `AddressAutofillOptions.Country` and `AddressAutofillOptions.Language` have been removed and replaced by `IsoCountryCode` and `IsoLanguageCode`.
+- [Autofill] `AddressAutofillResponse` type has been removed. Now `AddressAutofill` functions return `Expected<Exception, List<AddressAutofillSuggestion>>`.
 
 ### Mapbox dependencies
 - Search Native SDK `0.66.0`